Admin View Provinces

+++ frontend/src/pages/ProvincesAdmin.tsx
@@ -4,6 +4,9 @@ import { ProvinceLabel } from "../components/ProvinceLabel";
 import { FaSearch } from 'react-icons/fa';
 import { ProvinceModal } from "../components/ProvinceModal";
 import "../styles/ProvincesCRUD.css";
+import { NavZone } from "../components/NavZone.tsx";
+import { FaPlus } from "react-icons/fa";
+
 
 
 export function ProvincesAdmin() {
@@ -16,11 +19,11 @@ export function ProvincesAdmin() {
     const [provinces, setProvinces] = useState<Province[]>([]);
     const [filteredProvinces, setFilteredProvinces] = useState<Province[]>([]);
     const [loading, setLoading] = useState(true);
-    const [error, setError] = useState(null);
-    const [newName, setNewName] = useState('');
+    const [error, setError] = useState<any>();
     const [searchTerm, setSearchTerm] = useState('');
     const [modalVisible, setModalVisible] = useState(false);
     const [selectedProvince, setSelectedProvince] = useState<Province | null>(null);
+    const [modalAction, setModalAction] = useState<'create' | 'edit'>('create');
 
     useEffect(() => {
         fetch('/api/provinces')
@@ -46,54 +49,67 @@ export function ProvincesAdmin() {
 
     }, [searchTerm, provinces]);
 
+    useEffect(() => {
+        setTimeout(() => {
+            setError("");
+        }, 11000);
+    }, [error]);
+
     if (loading) {
         return <div>Loading...</div>;
     }
 //Falta estilizar cacheado de errores
-    if (error) {
-        return <div>Error: {error}</div>;
-    }
 
-
-
-    function addProvince() {
-        if (!newName.trim()) return;
+    function addProvince(nameProvince: string) {
+        if (!nameProvince.trim()) return;
         fetch('/api/provinces', {
             method: 'POST',
             headers: { 'Content-Type': 'application/json' },
-            body: JSON.stringify({ nameProvince: newName }) // idProvince se genera en el backend
+            body: JSON.stringify({ nameProvince: nameProvince }) // idProvince se genera en el backend
     })
       .then(res => {
-        if(res.ok) {
-          return res.json();
+        if(!res.ok) {
+          setModalVisible(false);
+          throw new Error('El nombre coincide con otra provincia existente');
         }
-        throw new Error('El nombre coincide con otra provincia existente');
+        return res.json();
+        
       })
       .then(created => {
         // añadimos la provincia nueva al array
         console.log('Provincia creada:', created);
+        setModalVisible(false);
+        setError({});
         setProvinces([...provinces, created.data]);
-        setNewName('');
       })
       .catch(err => {
-        alert('Error al crear provincia: ' + err.message);
+        setError({ baseMessage: "Error al crear provincia: ",
+                  detail: err.message
+        });
       });
   };
 
   function deleteProvince(id: string) {
+    if (!id) return;
+
     fetch(`/api/provinces/${id}`, {
       method: 'DELETE',
     })
       .then(res => {
-        if(!res.ok) throw new Error('Error al eliminar provincia');
+        if(!res.ok) {
+          setModalVisible(false);
+          throw new Error('Pruebe más tarde');}
       })
       .then(() => {
         // Eliminamos la provincia del array
         setProvinces(provinces.filter(prov => prov.idProvince !== id));
+        setError({});
         setModalVisible(false);
       })
       .catch(err => {
-        alert('Error al eliminar provincia: ' + err.message);
+        setError({ baseMessage: "Error al eliminar provincia: ",
+          detail: err.message
+        });
       });
   }
 
@@ -109,35 +125,27 @@ export function ProvincesAdmin() {
         if(res.ok) {
           return res.json();
         }
+        setModalVisible(false);
         throw new Error('El nombre de la provincia coincide con otra existente');
       })
       .then(updated => {
         const newProvinces = [updated.data, ...provinces.filter(prov => prov.idProvince !== id)];
         setProvinces(newProvinces);
+        setError({});
         setModalVisible(false);
       })
       .catch(err => {
-        alert('Error al modificar provincia: ' + err.message);
+        setError({ baseMessage: "Error al modificar provincia: ",
+          detail: err.message
+        });
       });
   }
 
   
-
-
     return (
       <>
-        {modalVisible && selectedProvince != null && (
-            <ProvinceModal
-                visible={modalVisible}
-                onClose={() => setModalVisible(false)}
-                province={selectedProvince}
-                onDelete={() => deleteProvince(selectedProvince.idProvince)}
-                onEdit={editProvince}
-            />
-        )
-        }
         <div className="admin-home">
-            <h1 style={{ textAlign: 'center' }}>Administrador de Provincias</h1>
+            <NavZone title="Administrador de Provincias"/>
             <div className="province-searchBar">
                 <FaSearch className="search-icon"/>
                 <input className="province-searchInput"
@@ -151,6 +159,7 @@ export function ProvincesAdmin() {
                     {filteredProvinces.map(prov => (
                     <li key={prov.idProvince} onClick={() => {
                       setModalVisible(true);
+                      setModalAction('edit');
                       setSelectedProvince(prov);
                     }}>
                         <ProvinceLabel name={prov.nameProvince} id={prov.idProvince} onDelete={() => deleteProvince(prov.idProvince)} onEdit={() => editProvince(prov.idProvince, "")} />
@@ -158,17 +167,34 @@ export function ProvincesAdmin() {
                 ))}
             </ul>
             </div>
-            
-              <h1>Crear Provincia</h1>
-            <div className="create-province">
-              <input className="create-province-input"
-                  type="text"
-                  placeholder="Nombre de provincia"
-                  value={newName}
-                  onChange={e => setNewName(e.target.value)}
-              />
-              <button onClick={addProvince}>Añadir</button>
-            </div>
+            <button className="add-button" onClick={() => (setModalVisible(true), setModalAction('create'))}><strong>Agregar Provincia </strong><FaPlus /></button>
+
+           {selectedProvince != null && modalAction === "edit" && (
+            <ProvinceModal
+                visible={modalVisible}
+                onClose={() => setModalVisible(false)}
+                province={selectedProvince}
+                onDelete={() => deleteProvince(selectedProvince.idProvince)}
+                onEdit={editProvince}
+                action={modalAction}
+                onCreate={() => {}}
+            />)}
+
+            {modalAction === "create" && (
+            <ProvinceModal
+                visible={modalVisible}
+                onClose={() => setModalVisible(false)}
+                province={null}
+                onCreate={addProvince}
+                onEdit={() => {}}
+                onDelete={() => {}}
+                action={modalAction}
+            />)}
+
+            {error != "" && (
+                <p className="error-message"><strong>{error.baseMessage}</strong>{error.detail}</p>
+            )}
+
         </div>
     </> 
     );
